+++ sfsimodels/__init__.py
@@ -12,6 +12,9 @@ from sfsimodels.models.foundations import FoundationPad, FoundationRaft
 from sfsimodels.models import material
 from sfsimodels.models.systems import SoilStructureSystem, TwoDSystem
 from sfsimodels.models.time import TimeSeries
+from sfsimodels.models.coordinates import Coords, GlobalCoords, PositionalCoords
+from sfsimodels.models.units import Units, GlobalUnits
+from sfsimodels.models.loads import Load, LoadAtCoords
 from sfsimodels.output import format_value, format_name, output_to_table
 from sfsimodels.files import ecp_dict_to_objects, load_json, loads_json, Output, migrate_ecp
 from sfsimodels.functions import clean_float, collect_serial_value, add_to_obj, interp_left